+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 
 ### Added
+- Added schemas for Question, QuestionType, QuestionCondition and VersionedQuestion
+- Added resolvers and models for Section, Tag and VersionedSection
 - Added acceptedTerms to User schema and to user sql table
 - Added schema for Section and Tag
 - Initial schema, model, mocks and resolver for Templates